Develop

+++ agrosupport/src/main/java/com/agrosupport/api/profile/domain/model/commands/UpdateAdvisorCommand.java
@@ -1,4 +1,6 @@
 package com.agrosupport.api.profile.domain.model.commands;
 
-public record UpdateAdvisorCommand(Long id, Integer rating) {
+import java.math.BigDecimal;
+
+public record UpdateAdvisorCommand(Long id, BigDecimal rating) {
 }
