+++ agrosupport/src/main/java/com/agrosupport/api/profile/interfaces/rest/resources/AdvisorResource.java
@@ -1,6 +1,8 @@
 package com.agrosupport.api.profile.interfaces.rest.resources;
 
+import java.math.BigDecimal;
+
 public record AdvisorResource(Long id,
                               Long userId,
-                              Integer rating) {
+                              BigDecimal rating) {
 }
